requirements.txt
pandas>=2.0.0
numpy>=1.26.0
scikit-learn>=1.4.0
xgboost>=2.0.0
joblib>=1.3.0
pyarrow>=14.0.0
shapely>=2.0.0
matplotlib>=3.8.0
streamlit>=1.35.0
ortools>=9.8.0

ezdxf>=1.1.0
tensorflow>=2.10.0
